+++ include/sealloc/bin.h
@@ -14,16 +14,16 @@ typedef struct run_state run_t;
 
 typedef struct bin_state {
   ll_head_t run_list_inactive;  // List of inactive runs
-  uint32_t
+  unsigned
       run_list_inactive_cnt;    // Number of runs currently in the inactive list
   run_t *runcur;                // Current active run servicing allocations
-  uint16_t reg_size;            // Size of single region
-  uint16_t run_size_pages;      // Number of pages that each run spans
-  uint32_t reg_mask_size_bits;  // Number of bits in region bitmask
+  unsigned reg_size;            // Size of single region
+  unsigned run_size_pages;      // Number of pages that each run spans
+  unsigned reg_mask_size_bits;  // Number of bits in region bitmask
 } bin_t;
 
 // Initialize bin
-void bin_init(bin_t *bin, uint16_t reg_size);
+void bin_init(bin_t *bin, unsigned reg_size);
 
 // Delete inactive run
 void bin_delete_run(bin_t *bin, run_t *run);
